+++ test/GetEvent_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/fingerprintjs/fingerprint-pro-server-api-go-sdk/config"
-	"github.com/fingerprintjs/fingerprint-pro-server-api-go-sdk/sdk"
+	"github.com/fingerprintjs/fingerprint-pro-server-api-go-sdk/v2/config"
+	"github.com/fingerprintjs/fingerprint-pro-server-api-go-sdk/v2/sdk"
 	"github.com/stretchr/testify/assert"
 	"log"
 	"net/http"
@@ -52,4 +52,88 @@ func TestReturnsEvent(t *testing.T) {
 	assert.NoError(t, err)
 	assert.NotNil(t, res)
 	assert.Equal(t, res, mockResponse)
+}
+
+func TestReturnsBotdTooManyRequestsError(t *testing.T) {
+	mockResponse := GetMockEventResponse("../test/mocks/get_event_botd_too_many_requests_error.json")
+
+	ts := httptest.NewServer(http.HandlerFunc(func(
+		w http.ResponseWriter,
+		r *http.Request,
+	) {
+		configFile := config.ReadConfig("../config.json")
+		integrationInfo := r.URL.Query().Get("ii")
+		assert.Equal(t, integrationInfo, fmt.Sprintf("fingerprint-pro-server-go-sdk/%s", configFile.PackageVersion))
+		assert.Equal(t, r.URL.Path, "/events/123")
+
+		apiKey := r.Header.Get("Auth-Api-Key")
+		assert.Equal(t, apiKey, "api_key")
+
+		w.Header().Set("Content-Type", "application/json")
+
+		err := json.NewEncoder(w).Encode(mockResponse)
+
+		if err != nil {
+			log.Fatal(err)
+		}
+	}))
+	defer ts.Close()
+
+	cfg := sdk.NewConfiguration()
+	cfg.ChangeBasePath(ts.URL)
+
+	client := sdk.NewAPIClient(cfg)
+
+	ctx := context.WithValue(context.Background(), sdk.ContextAPIKey, sdk.APIKey{
+		Key: "api_key",
+	})
+
+	res, _, err := client.FingerprintApi.GetEvent(ctx, "123")
+
+	assert.NoError(t, err)
+	assert.NotNil(t, res)
+	assert.Equal(t, res, mockResponse)
+	assert.Equal(t, res.Products.Botd.Error_.Code, "TooManyRequests")
+}
+
+func TestReturnsIdentificationTooManyRequestsError(t *testing.T) {
+	mockResponse := GetMockEventResponse("../test/mocks/get_event_identification_too_many_requests_error.json")
+
+	ts := httptest.NewServer(http.HandlerFunc(func(
+		w http.ResponseWriter,
+		r *http.Request,
+	) {
+		configFile := config.ReadConfig("../config.json")
+		integrationInfo := r.URL.Query().Get("ii")
+		assert.Equal(t, integrationInfo, fmt.Sprintf("fingerprint-pro-server-go-sdk/%s", configFile.PackageVersion))
+		assert.Equal(t, r.URL.Path, "/events/123")
+
+		apiKey := r.Header.Get("Auth-Api-Key")
+		assert.Equal(t, apiKey, "api_key")
+
+		w.Header().Set("Content-Type", "application/json")
+
+		err := json.NewEncoder(w).Encode(mockResponse)
+
+		if err != nil {
+			log.Fatal(err)
+		}
+	}))
+	defer ts.Close()
+
+	cfg := sdk.NewConfiguration()
+	cfg.ChangeBasePath(ts.URL)
+
+	client := sdk.NewAPIClient(cfg)
+
+	ctx := context.WithValue(context.Background(), sdk.ContextAPIKey, sdk.APIKey{
+		Key: "api_key",
+	})
+
+	res, _, err := client.FingerprintApi.GetEvent(ctx, "123")
+
+	assert.NoError(t, err)
+	assert.NotNil(t, res)
+	assert.Equal(t, res, mockResponse)
+	assert.Equal(t, res.Products.Identification.Error_.Code, "429 Too Many Requests")
 }