+++ test/utils.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/fingerprintjs/fingerprint-pro-server-api-go-sdk/sdk"
+	"github.com/fingerprintjs/fingerprint-pro-server-api-go-sdk/v2/sdk"
 )
 
 func readFromFileAndUnmarshal(path string, i interface{}) {
@@ -24,12 +24,12 @@ func readFromFileAndUnmarshal(path string, i interface{}) {
 
 func GetMockResponse(path string) sdk.Response {
 	var mockResponse sdk.Response
-	readFromFileAndUnmarshal(path, mockResponse)
+	readFromFileAndUnmarshal(path, &mockResponse)
 	return mockResponse
 }
 
 func GetMockEventResponse(path string) sdk.EventResponse {
 	var mockResponse sdk.EventResponse
-	readFromFileAndUnmarshal(path, mockResponse)
+	readFromFileAndUnmarshal(path, &mockResponse)
 	return mockResponse
 }